dosbox-x: import dosbox-x-0.82.7 as wip/dosbox-x
As a fork of DOSBox, it retains compatibility with the wide base
of DOS games and DOS gaming DOSBox was designed for. DOSBox-X
however goes further, with a focus on accurate emulation of the
hardware and many more ways to tweak and configure the DOS virtual
machine! We believe that a better way to emulate the legacy PC
platform is to give the user all the options they need to emulate
everything from original IBM PC hardware with 64KB of RAM all the
way up to late 90's hardware, whatever it takes to get that game
or software package to run. Our goal is to eventually make DOSBox-X
a complete emulation package that covers all pre-2000 DOS and
Windows 9x based hardware scenarios, including peripherals,
motherboards, CPUs, and all manner of hardware that was made for
PC hardware of that time.