go-ghodss-yaml: Import go-ghodss-yaml-1.0.0 as wip/go-ghodss-yaml
A wrapper around go-yaml designed to enable a better way of handling
YAML when marshaling to and from structs.

In short, this library first converts YAML to JSON using go-yaml and
then uses json.Marshal and json.Unmarshal to convert to or from the
struct. This means that it effectively reuses the JSON struct tags as well
as the custom JSON methods MarshalJSON and UnmarshalJSON unlike go-yaml.