crystal-bin: Import version 0.25.0
Crystal is a programming language with the goals of having a syntax
similar to Ruby (but compatibility with it is not a goal), being
statically type-checked, but without having to specify the type of
variables or method arguments, being able to call C code by writing
bindings to it in Crystal, having compile-time evaluation and
generation of code, to avoid boilerplate code, and compiling to
efficient native code.

Currently this package installs a binary for Darwin/x86_64 only.