fonttosfnt: update to 1.0.5.
Is it a bird? is it a plane? No it's a fonttofs.. fontstof.. fonttofn...
it's a release of the tool that wraps a bitmap font into a truetype wrapper!
Bringing you the last 9 years of exciting changes including ansification and
lots of autotools cleanup changes, because that is how we rolled 7 years
ago. (Almost) no source files were harmed in the making of this release.