tmux: Import tmux-2.8rc as wip/tmux
Tmux is a "terminal multiplexer", it enables a number of terminals
(or windows) to be accessed and controlled from a single terminal.
Tmux is intended to be a simple, modern, BSD-licensed alternative
to programs such as GNU screen.

Based on misc/tmux.
This package is intended to ease testing to update misc/tmux when tmux-2.8
will be out.