webkit-gtk: Import webkitgtk-2.20.5nb1 as wip/webkit-gtk
WebKit is an open source web browser engine. WebKit is also the name of
the Mac OS X system framework version of the engine that's used by
Safari, Dashboard, Mail, and many other OS X applications. WebKit's HTML
and JavaScript code began as a branch of the KHTML and KJS libraries
from KDE.

This is the GTK port of the engine.

This is the same version at the moment in www/webkit-gtk, it will
be soon updated to 2.22.2 and then further patch to add initial
JIT support for NetBSD will be committed.