zathura: Import zathura-0.4.1 as wip/zathura
zathura is a highly customizable and functional PDF viewer based
on the poppler rendering library and the gtk+ toolkit. The idea
behind zathura is an application that provides a minimalistic and
space saving interface as well as an easy usage that mainly focuses
on keyboard interaction.

Based on print/zathura and intended to be used to update it.