Add not pacman 1.4
Ever wondered how Pacman would be like with physics? No? Well, it's
revolutionary. This mashup of "Not" and "Pacman" puts The Pac and
the Ghosts in a in a Pacworld cage, at mercy of gravity and the
level's shape, controlled solely by you. Ghosts still don't like
you very much, so you should be careful to make Pacman avoid them.