unittest-cpp: Import unittest-cpp-2.0.0 as wip/unittest-cpp
UnitTest++ is a lightweight unit testing framework for C++. It was
designed to do test-driven development on a wide variety of platforms.
Simplicity, portability, speed, and small footprint are all very
important aspects of UnitTest++. UnitTest++ is mostly standard C++
and makes minimal use of advanced library and language features,
which means it should be easily portable to just about any platform.