chicken5: Update wip/chicken5 to 5.0.0rc4
Changes:
5.0.0rc4
--------
This release candidate fixes a handful of issues with chicken-install
and chicken-do.  For an overview of fixed tickets since 5.0.0rc3, see
the Trac page:

https://bugs.call-cc.org/query?status=closed&version=5.0.0rc3&milestone=5.0&group=status&col=id&col=summary&col=owner&col=type&col=priority&col=component&col=version&order=priority

Given that there were only a few issues found in rc3, it's likely that
this will be the last release candidate so we can make an official
release soon!