Qemu with NVMM acceleration support.
	qemu-system-x86_64 [... whatever ...] -accel nvmm

Same as the 'qemu' package, but with patch-nvmm-support applied, and only
x86_64 as target.

Builds, launches, but doesn't work yet. The second call to nvmm_gpa_map fails,
because qemu is trying to map twice the same gpa, and I don't understand what
is the expected behavior in this case. I guess I'll have to ask them.

Eventually the patch will be merged upstream, and our generic 'qemu' package
will be built with --enable-nvmm by default on NetBSD.