Update autossh to 1.4f.
For Mayuresh.

Update: Version 1.4f incorporates a number of changes and fixes to signal
handling and exit conditions. The most notable changes are that it will exit
properly when killed while waiting between ssh restarts, and it will now
restart rather than exiting when the child ssh process is killed.