Import OpenJazz 20180522 as wip/openjazz.
OpenJazz is a free, open-source version of the classic Jazz Jackrabbit™
games.

OpenJazz can be compiled on a wide range of operating systems, including
Windows, macOS, GNU/Linux and *BSD. Also ports are available for some
homebrew platforms, for example Wii and PSP.

To play, you will need the files from one of the original games.