openfodder-data: import openfodder-data-1.5 as wip/openfodder-data
Cannon Fodder is an action-strategy shoot 'em up game developed by
Sensible Software and published by Virgin Interactive.  The game
is military-themed and based on shooting action but with a strategy
game-style control system. The player directs troops through numerous
missions, battling enemy infantry, vehicles and installations.

This is the data for the engine wip/openfodder.