Import bcal
bcal (Byte CALculator) is a REPL CLI utility for storage expressions,
unit conversions or address calculations. If you can't calculate the
hex address offset for (512 - 16) MiB, or the value when the 43rd bit
of a 64-bit address is set mentally, bcal is for you.

It has a bc(1) mode for general-purpose numerical calculations.

Only 64-bit operating systems are supported.