libdazzle: Import libdazzle-3.30.2 as wip/libdazzle
The libdazzle library is a companion library to GObject and Gtk+.
It provides various features that we wish were in the underlying
library but cannot for various reasons.
In most cases, they are wildly out of scope for those libraries.
In other cases, our design isn't quite generic enough to work for everyone.

TODO: Fails to link with:
TODO: `ERROR: can't resolve libraries to shared libraries: dazzle-1.0'
TODO: For complete error log please give a look to the TODO file.