tex-listofitems: Imported 1.6
listofitems - Grab items in lists using user-specified sep char

   This simple package is designed to read a list of items whose parsing
character may be selected by the user.
   Once the list is read, its items are stored in a structure that behaves
as a dimensioned array. As such, it becomes very easy to access an item in
the list by its number.
   For example, if the list is stored in the macro \foo, the item #3 is
designated by \foo[3].
   A component may, in turn, be a list with a parsing delimiter different
from the parent list, paving the way for nesting and employing a syntax
reminiscent of an array of several dimensions of the type \foo[3,2] to
access the item #2 of the list contained within the item #3 of the top-tier
list.

Changelog for 1.6:
--------------------------------------------------------------------------------
- \defpair defines paired tokens in items
--------------------------------------------------------------------------------