guile-bytestructures: Import guile-bytestructures 1.0.5 as wip/guile-bytestructures
guile-bytestructures is a library offering a system imitating the type system of the
C programming language, to be used on bytevectors.
C's type system works on raw memory, and Scheme ones work on bytevectors
which are an abstraction over raw memory in Scheme.
The system elevates types to first-class status.