ircd-hybrid: Remove reference.conf from CONF_FILES_PERMS
doc/reference.conf seems a reference example for a possible ircd.conf and it
seems not used. Do not treat it as a configuration file.

XXX: patches/patch-doc_Makefile.in probably needs to be adjusted in order to
XXX: just do not do anything because doc/* files are already handled in pkgsrc
XXX: post-install: target.