swi-prolog-packages: Import swi-prolog-packages-8.0.1 as wip/swi-prolog-packages
Packages for TCP/IP, XML/SGML, RDF, and the GUI toolkit XPCE for
swi-prolog, an ISO/Edinburgh-style Prolog compiler.

Based on lang/swi-prolog-packages and intended to be used to update it...
Still a lot of work TODO though, please give a look to the TODO for more
information! (and XXX comments!)