wip/lld-netbsd: import lld-9.0.0nb20190201
LLD is a linker from the LLVM project that is a drop-in replacement for system
linkers and runs much faster than them. It also provides features that are
useful for toolchain developers.

The linker supports ELF (Unix), PE/COFF (Windows), Mach-O (macOS) and
WebAssembly in descending order of completeness. Internally, LLD consists of
several different linkers. The ELF port is the one that will be described in
this document. The PE/COFF port is complete, including Windows debug info (PDB)
support. The WebAssembly port is still a work in progress (See WebAssembly lld
port). The Mach-O port is built based on a different architecture than
the others. For the details about Mach-O, please read ATOM-based lld.