ats2: Update to 0.3.13
0.3.13

This is the 50th release of ATS2, the successor of the
ATS programming language. The compiler for ATS2 is given
the name ATS/Positats, ATS2/Postiats or simply Postiats.

The official website for ATS is:

http://www.ats-lang.org

ATS-Postiats is hosted at github:

https://github.com/githwxi/ATS-Postiats

Major releases of ATS2 are available at:

https://sourceforge.net/projects/ats2-lang/

Major releases of external packages for ATS2 are available at:

https://sourceforge.net/projects/ats2-lang-contrib/

Here is a list of major additions and changes since the last release:

1. Fixing the issue of escaping '$' in PHP string literals (atscc2php)
   Kudos to M88!
2. Fixing very minor issues in libats/qlist and libats/linmap
3. Enforcing that no d2var can be leaked/exported from a namespace