lapack-ng: experimental new BLAS+LAPACK package
This is a step towards the scheme of choice of BLAS and LAPACK
libraries from multiple providers and C wrappers (cblas and lapacke)
on top. It builds both BLAS and LAPACK libraries from Netlib using
the new cmake-based build. This simplifies the package as no hacking
is needed for shared libs.

More work to follow in the coming days.