go-gomail: Import go-gomail-2.0.0+ as wip/go-gomail
Gomail is a simple and efficient package to send emails.
It is well tested and documented.

Gomail can only send emails using an SMTP server. But the API is flexible
and it is easy to implement other methods for sending emails using a local
Postfix, an API, etc.

This is newer than version 2.0.0 as mailsend-go requires it.