Add sauerbraten.
Mostly seems to work (besides configuration not being saved and
there being no application entries). Everyone else is shipping an
older version though, I wonder if there's a reason or if they're
just lazy.

Cube 2: Sauerbraten is a free multiplayer & singleplayer first
person shooter, the successor of the Cube FPS.

Much like the original Cube, the aim of this game is fun, old school
deathmatch gameplay and also to allow map/geometry editing to be
done cooperatively in-game.

The engine supporting the game is entirely original in code &
design, and its code is Open Source (ZLIB license, read the docs
for more on how you can use the engine).