XNedit: Import xnedit-1.2.3 as wip/xnedit
XNedit is a fast and classic X11 text editor, based on NEdit, with
full unicode support and antialiased text rendering.

NEdit is a standard GUI (Graphical User Interface) style text editor
for programs and plain-text files.  Users of Macintosh and MS
Windows based text editors should find NEdit a familiar and
comfortable environment.  NEdit provides all of the standard menu,
dialog, editing, and mouse support, as well as all of the standard
shortcuts to which the users of modern GUI based environments are
accustomed.  For users of older style Unix editors, welcome to the
world of mouse-based editing!