wip/michabo: Import as michabo-d37d39e5a43fef036d456b5f0f76410a6f8f5397
Michabo is a work in progress native (e.g. not Electron) desktop app
for Pleroma and Mastodon servers.
Warning! This app is not yet suitable for use as your daily driver.
While the client core is robust, there are many missing features which
make it annoying to use.

NOTE: this still crashes.