zoneminder: Address thread portability and type issues
zoneminder seems to on one hand assume pthreads and the other not.  On
NetBSD and Solaris, assume pthreads harder.

Cast tv_use to long, because the code uses %ld, and that's easier than
trying to change more.  (This update will be superceded anyway, and
the time type issues need to be fixed upstream.)