iosevka-cc-slab-ttf: Import iosevka-cc-slab-ttf-2.2.1 as wip/iosevka-cc-slab-ttf
Iosevka is a slender monospace sans-serif and slab-serif typeface inspired
by Pragmata Pro, M+ and PF DIN Mono, designed to be the ideal font for
programming.

This package is "CC Slab" variant.

Some of the symbols are en-widen to maintain compatibility with CJK
typefaces. Ligatures are included.

The font is slab-serif.