tiff-32: compat32 package for tiff
This is software provides support for the Tag Image Format (TIFF).
Included is a library, libtiff, for reading and writing TIFF, a
collection of tools for doing simple manipulations of TIFF images, and
documentation on the library and tools.  A random assortment of
TIFF-related software contributed by others is also included.

The library, along with associated tool programs, should handle most
of your needs for reading and writing files that follow the 5.0 or 6.0
TIFF spec.  There is also considerable support for some of the more
esoteric portions of the 6.0 TIFF spec.

This package is intended to crosscompile tiff for i386.