libxslt-32: compat32 package for libxslt
Libxslt is the XSLT C library developed for the Gnome project. XSLT itself is
an XML language to define transformation for XML. Libxslt is based on libxml2
the XML C library developed for the Gnome project. It also implements most of
the EXSLT set of extensions functions and some of Saxon's evaluate and
expressions extensions.

This package is intended to crosscompile libxslt for i386.