Add arcan 0.5.5.2.
Arcan is a powerful development framework for creating virtually
anything from user interfaces for specialized embedded applications
all the way to full-blown standalone desktop environments.

At its heart lies a robust and portable multimedia engine, with a
well-tested and well-documented Lua scripting interface. The
development emphasizes security, debuggability and performance --
guided by a principle of least surprise in terms of API design.

Adapted to netbsd, but it doesn't run on it yet. Note that in TODO.