mosquitto: Rototill config file patch
Rather than trying to not install the example config (which is not
named as an example), let it be, and add a section where upstream
choices inconsistent with traditional UNIX norms are adjusted, and add
a section for users of the package to put their changes.

Default to the openssl cert dir for ca_path, in the hopes that
mosquitto will use the system's configured trust anchor set.