alacritty: import as wip/alacritty
This does not build yet and tracks a commit and PR working on
NetBSD support.

Alacritty is a terminal emulator with a strong focus on simplicity and
performance. With such a strong focus on performance, included
features are carefully considered and you can always expect Alacritty
to be blazingly fast. By making good choices for defaults, Alacritty
requires no additional setup. However, it does allow configuration of
many aspects of the terminal.

The software is considered to be at a beta level of readiness -- there
are a few missing features and bugs to be fixed, but it is already
used by many as a daily driver.