py-weewx: Add version 3.8.2
WeeWX is a free, open source, software program, written in Python,
which interacts with your weather station to produce graphs, reports,
and HTML pages. It can optionally publish to weather sites or web
servers. It uses modern software concepts, making it simple, robust,
and easy to extend. It includes extensive documentation.

WeeWX runs under most versions of Linux, as well as macOS, *BSD, and
Solaris. Many users are running on the Raspberry Pi. The images on
this page and throughout this web site are from sample stations
running WeeWX.

Thousands of stations throughout the world run WeeWX, many of whom
have opted-in to be shown on our station map.

pkgsrc note: For now, this package is useful to install the
dependencies; see the Makefile for configuration issues.