add velox version 0.0.20190810
velox is a simple window manager based on swc. It is inspired by
dwm and xmonad.

velox uses tag-based window management similar to dwm. This allows
you to construct your workspace by selecting a set of tags to work
with. However, in contrast with dwm, windows do not have any screen
associated with them; they are shown on whichever screen has their
tag selected, allowing you to easily move windows between screens
by selecting their tag on a different screen. This is similar to
the multi-monitor workspace switching in xmonad.

To ensure that we never attempt to show a window in two places at
once, we have to impose several constraints. First, each window
must have exactly one tag. In practice, I've found that I rarely
intentionally mark a window with more than one tag anyway. Second,
when you select a tag that is currently display on a different
screen, the tag is first deselected from that screen.