doom2-pwad-preacher: Import as wip/doom2-pwad-preacher
Preacher is a 9 map boom compatible episode. You play
as a mad preacher who hears a voice in his head
impelling him to take down all evil in the World.

Layouts are open and non linear, and start spots are
randomized.

There are graphic and sound replacements, but weapon
and enemy behaviours are untouched.